server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
pytest==6.2.5
pytest-cov==2.12.0
pytest-mock==3.6.1
black==21.6b0
pylint==2.8.2 